added public api info added check for entry existence - we don't want to rewrite the file if no item is removed

+++ database/database.php
@@ -1,4 +1,23 @@
 <?php
+/*
+ * DataBase class to handle the data
+ *
+ *          PUBLIC API
+ * ================================
+ * If no such table exists, it creates new with name equal to the first parameter!
+ * The input must be associative array or object for single entry insertion!
+ * In case you need to add multiple items, pass them as array of associative arrays and/or objects!
+ * function addToTable( tableName, input );
+ * ================================
+ * Returns array with the entries in the table with name equal to the first parameter
+ * The entries are objects
+ * function getTableContents( tableName );
+ * ================================
+ * Removes element from the table with id equal to the second parameter
+ * If no such element exists, no action is taken!
+ * function removeItemById( $tableName, $id );
+ * ================================
+ */
 class DataBase
 {
     private $dataPath;
@@ -44,14 +63,20 @@ class DataBase
     {
         $tableData = $this->getTableContents( $tableName );
         $filtered = [];
+        $entryExists = false;
+
         foreach ($tableData as $entry) {
             if ($entry->id != $id) {
                 $filtered[] = $entry;
+            } else {
+                $entryExists = true;
             }
         }
 
-        $filePath = $this->dataPath . $tableName;
-        file_put_contents( $filePath, json_encode( $filtered ), FILE_USE_INCLUDE_PATH );
+        if ($entryExists) {
+            $filePath = $this->dataPath . $tableName;
+            file_put_contents( $filePath, json_encode( $filtered ), FILE_USE_INCLUDE_PATH );
+        }
     }
 
     private function isAssociative( $array )
